+++ src/handlers/mod.rs
@@ -2,6 +2,8 @@ pub mod config;
 pub mod descriptor;
 #[cfg(feature = "dns_payment")]
 pub mod dns;
+#[cfg(feature = "hwi")]
+pub mod hwi;
 pub mod key;
 pub mod offline;
 pub mod online;
@@ -114,7 +116,8 @@ pub trait AppCommand<C> {
     feature = "esplora",
     feature = "rpc",
     feature = "cbf",
-    feature = "dns_payment"
+    feature = "dns_payment",
+    feature = "hwi"
 ))]
 pub trait AsyncAppCommand<C> {
     type Output: FormatOutput;